refactor: :zap: Agregado el manejo de la relacion M:M usando JPA
Se eliminan todas las clases relacionadas a la tabla subCategoriaSolicitud
