+++ backend/dtos/src/main/java/mercadeoucab/dtos/DtoSolicitud.java
@@ -11,7 +11,7 @@ public class DtoSolicitud extends DtoBase{
     private DtoUsuario usuario;
     private DtoMarca marca;
     private DtoTipo tipo;
-    private List<DtoTipo> tipos;
+    private DtoSubCategoria subCategoria;
 
     public DtoSolicitud(long id)throws Exception{ super(id); }
     public DtoSolicitud(){}
@@ -25,18 +25,17 @@ public class DtoSolicitud extends DtoBase{
     public DtoUsuario getUsuario() {return usuario;}
     public void setUsuario(DtoUsuario usuario) {this.usuario = usuario;}
 
-    public List<DtoTipo> getTipos() {
-        return tipos;
-    }
-    public void setTipos(List<DtoTipo> tipos) {
-        this.tipos = tipos;
-    }
-
     public DtoTipo getTipo() {
         return tipo;
     }
-
     public void setTipo(DtoTipo tipo) {
         this.tipo = tipo;
+    }
+
+    public DtoSubCategoria getSubCategoria() {
+        return subCategoria;
+    }
+    public void setSubCategoria(DtoSubCategoria subCategoria) {
+        this.subCategoria = subCategoria;
     }
 }
