+++ backend/servicio/src/main/java/mercadeoucab/servicio/ServicioSolicitud.java
@@ -3,10 +3,7 @@ package mercadeoucab.servicio;
 import mercadeoucab.accesodatos.DaoSolicitud;
 import mercadeoucab.accesodatos.DaoTipo;
 import mercadeoucab.dtos.DtoSolicitud;
-import mercadeoucab.entidades.Marca;
-import mercadeoucab.entidades.Solicitud;
-import mercadeoucab.entidades.Tipo;
-import mercadeoucab.entidades.Usuario;
+import mercadeoucab.entidades.*;
 
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
@@ -67,6 +64,8 @@ public class ServicioSolicitud extends AplicacionBase{
             solicitud.setMarca( marca );
             Tipo tipo = new Tipo(dtoSolicitud.getTipo().get_id());
             solicitud.addTipo(tipo);
+            SubCategoria subCategoria = new SubCategoria(dtoSolicitud.getSubCategoria().get_id());
+            solicitud.addSubCategoria(subCategoria);
             Solicitud resul = dao.insert( solicitud );
             resultado.set_id( resul.get_id());
         }catch (Exception e) {
